+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added "Expore" link to admin panel.
 - Added "Profile directory", "Custom feeds" and "Groups" to "Explore" page.
 - Included profile menu action for finding custom feeds to which the user has been added.
+- Added "Mute conversation" option to post menu.
 
 ### Changed